+++ 3rdparty/zint-2.6.1/backend/gs1.c
@@ -77,7 +77,7 @@ int gs1_verify(struct zint_symbol *symbol, const unsigned char source[], const s
     int error_latch;
 
     /* Detect extended ASCII characters */
-    for (i = 0; i < src_len; i++) {
+    for (i = 0; i < (int)src_len; i++) {
         if (source[i] >= 128) {
             strcpy(symbol->errtxt, "250: Extended ASCII characters are not supported by GS1");
             return ZINT_ERROR_INVALID_DATA;
@@ -101,7 +101,7 @@ int gs1_verify(struct zint_symbol *symbol, const unsigned char source[], const s
     min_ai_length = 5;
     j = 0;
     ai_latch = 0;
-    for (i = 0; i < src_len; i++) {
+    for (i = 0; i < (int)src_len; i++) {
         ai_length += j;
         if (((j == 1) && (source[i] != ']')) && ((source[i] < '0') || (source[i] > '9'))) {
             ai_latch = 1;
@@ -158,7 +158,7 @@ int gs1_verify(struct zint_symbol *symbol, const unsigned char source[], const s
     }
 
     ai_count = 0;
-    for (i = 1; i < src_len; i++) {
+    for (i = 1; i < (int)src_len; i++) {
         if (source[i - 1] == '[') {
             ai_location[ai_count] = i;
             j = 0;
@@ -183,7 +183,7 @@ int gs1_verify(struct zint_symbol *symbol, const unsigned char source[], const s
         data_length[i] = 0;
         do {
             data_length[i]++;
-        } while ((source[data_location[i] + data_length[i] - 1] != '[') && (data_location[i] + data_length[i] <= src_len));
+        } while ((source[data_location[i] + data_length[i] - 1] != '[') && (data_location[i] + data_length[i] <= (int)src_len));
         data_length[i]--;
     }
 
@@ -301,7 +301,7 @@ int gs1_verify(struct zint_symbol *symbol, const unsigned char source[], const s
     j = 0;
     last_ai = 0;
     ai_latch = 1;
-    for (i = 0; i < src_len; i++) {
+    for (i = 0; i < (int)src_len; i++) {
         if ((source[i] != '[') && (source[i] != ']')) {
             reduced[j++] = source[i];
         }
