+++ 3rdparty/zint-2.6.1/backend/imail.c
@@ -274,7 +274,7 @@ int imail(struct zint_symbol *symbol, unsigned char source[], size_t length) {
 
     read = 0;
     j = 0;
-    for (i = 0; i < length; i++) {
+    for (i = 0; i < (int)length; i++) {
         if (source[i] == '-') {
             tracker[read] = '\0';
             j = 1;
@@ -314,7 +314,7 @@ int imail(struct zint_symbol *symbol, unsigned char source[], size_t length) {
         accum[i] = 0;
     }
 
-    for (read = 0; read < strlen(zip); read++) {
+    for (read = 0; read < (int)strlen(zip); read++) {
 
         for (i = 0; i < 112; i++) {
             x_reg[i] = accum[i];
@@ -359,7 +359,7 @@ int imail(struct zint_symbol *symbol, unsigned char source[], size_t length) {
         accum[i] = 0;
     }
 
-    for (read = 0; read < strlen(zip_adder); read++) {
+    for (read = 0; read < (int)strlen(zip_adder); read++) {
 
         for (i = 0; i < 112; i++) {
             y_reg[i] = accum[i];
@@ -426,7 +426,7 @@ int imail(struct zint_symbol *symbol, unsigned char source[], size_t length) {
 
     /* and then the rest */
 
-    for (read = 2; read < strlen(tracker); read++) {
+    for (read = 2; read < (int)strlen(tracker); read++) {
 
         for (i = 0; i < 112; i++) {
             y_reg[i] = accum[i];
@@ -584,7 +584,7 @@ int imail(struct zint_symbol *symbol, unsigned char source[], size_t length) {
 
     /* Translate 4-state data pattern to symbol */
     read = 0;
-    for (i = 0; i < strlen(data_pattern); i++) {
+    for (i = 0; i < (int)strlen(data_pattern); i++) {
         if ((data_pattern[i] == '1') || (data_pattern[i] == '0')) {
             set_module(symbol, 0, read);
         }
